Moved RestUtils to twc module
