+++ twc/src/main/java/org/openmbee/sdvc/twc/utilities/TwcPermissionUtils.java
@@ -4,7 +4,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import org.openmbee.sdvc.core.utils.RestUtils;
 import org.openmbee.sdvc.twc.TeamworkCloud;
 import org.openmbee.sdvc.twc.TeamworkCloudEndpoints;
 import org.openmbee.sdvc.twc.constants.TwcConstants;
